+++ code/app/src/main/java/com/nicholasrutherford/distractme/helpers/tasks/DatabaseTask.kt
@@ -1,11 +1,10 @@
-package com.nicholasrutherford.distractme.helpers
+package com.nicholasrutherford.distractme.helpers.tasks
 
 import android.os.AsyncTask
-import com.nicholasrutherford.distractme.activitys.MainActivity
 import com.nicholasrutherford.distractme.data.room.SavedArticlesDatabase
 import com.nicholasrutherford.distractme.data.room.SavedArticlesEntity
 
-class DatabaseTask(var activity: MainActivity, var db: SavedArticlesDatabase?, private val savedArticleList: ArrayList<SavedArticlesEntity>) : AsyncTask<Void, Void, Void>() {
+class DatabaseTask(var db: SavedArticlesDatabase?, private val savedArticleList: ArrayList<SavedArticlesEntity>) : AsyncTask<Void, Void, Void>() {
 
     override fun doInBackground(vararg params: Void?): Void? {
         savedArticleList.clear()
