+++ code/app/src/main/res/layout/article_home_layout.xml
@@ -18,38 +18,47 @@
 
     <TextView
         android:id="@+id/tvArticleTitle"
-        android:contentDescription="@string/title_article_desc"
-        android:ems="8"
-        android:layout_width="wrap_content"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/default_margin_or_padding"
+        android:contentDescription="@string/title_article_desc"
         android:text="@string/app_name"
         android:textSize="@dimen/default_header_tv_size"
-        android:layout_marginStart="@dimen/default_margin_or_padding"
-        android:layout_marginLeft="@dimen/default_margin_or_padding"
-        android:layout_marginTop="@dimen/default_margin_or_padding"
+        android:paddingLeft="@dimen/default_margin_or_padding"
+        android:paddingStart="@dimen/default_margin_or_padding"
+        tools:ignore="RtlSymmetry"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintStart_toEndOf="@+id/ivArticle"
         app:layout_constraintTop_toTopOf="@+id/ivArticle" />
 
     <TextView
         android:id="@+id/tvArticleDesc"
         android:contentDescription="@string/desc_article_text"
-        android:layout_width="wrap_content"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/default_margin_or_padding"
-        android:ems="12"
         android:textSize="@dimen/default_body_tv_size"
         android:text="@string/latin_text_filler"
+        android:paddingLeft="@dimen/default_margin_or_padding"
+        android:paddingStart="@dimen/default_margin_or_padding"
+        tools:ignore="RtlSymmetry"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintStart_toStartOf="@+id/tvArticleTitle"
         app:layout_constraintTop_toBottomOf="@+id/tvArticleTitle" />
 
     <TextView
         android:id="@+id/tvAuthorTitle"
-        android:layout_width="wrap_content"
+        android:layout_width="0sp"
         android:layout_height="wrap_content"
         android:textSize="@dimen/default_author_tv_size"
         android:textAlignment="center"
-        android:ems="8"
         android:layout_marginTop="@dimen/default_margin_or_padding"
+        android:paddingLeft="@dimen/default_margin_or_padding"
+        android:paddingStart="@dimen/default_margin_or_padding"
+        tools:ignore="RtlSymmetry"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintEnd_toEndOf="@+id/ivArticle"
         app:layout_constraintStart_toStartOf="@+id/ivArticle"
         app:layout_constraintTop_toBottomOf="@+id/ivArticle"
@@ -78,19 +87,27 @@
 
     <TextView
         android:id="@+id/tvSource"
-        android:layout_width="wrap_content"
+        android:layout_width="0sp"
         android:layout_height="wrap_content"
-        android:ems="11"
         android:layout_marginTop="@dimen/default_margin_or_padding"
+        android:paddingLeft="@dimen/default_margin_or_padding"
+        android:paddingStart="@dimen/default_margin_or_padding"
+        tools:ignore="RtlSymmetry"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintStart_toStartOf="@+id/tvArticleDesc"
         app:layout_constraintTop_toBottomOf="@+id/tvArticleDesc" />
 
     <TextView
         android:id="@+id/tvArticleDate"
-        android:layout_width="wrap_content"
+        android:layout_width="0sp"
         android:layout_height="wrap_content"
         android:layout_marginTop="5sp"
-        android:ems="13"
+        android:paddingLeft="@dimen/default_margin_or_padding"
+        android:paddingStart="@dimen/default_margin_or_padding"
+        tools:ignore="RtlSymmetry"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintStart_toStartOf="@+id/tvSource"
         app:layout_constraintTop_toBottomOf="@+id/tvSource" />
